+++ packages/swapper/src/swappers/zrx/utils.ts
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import BigNumber from 'bignumber.js'
 
 const axiosConfig = {
   baseURL: 'https://api.0x.org/',
@@ -10,3 +11,8 @@ const axiosConfig = {
 }
 
 export const zrxService = axios.create(axiosConfig)
+
+export const normalizeAmount = (amount: string | undefined): string | undefined => {
+  if (!amount) return undefined
+  return new BigNumber(amount).toNumber().toLocaleString('fullwide', { useGrouping: false })
+}
